pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid_edge"
version = "0.1.0"
description = "Secure Hybrid Edge-to-Cloud AI Tunneling"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "flask",
    "python-dotenv",
    "streamlit"
]

[project.scripts]
# These become terminal commands!
hybrid-server = "hybrid_edge.server:main"
hybrid-client = "hybrid_edge.client:main"